feat(extensibility): Add hint how to get access to a command from extension
In case user executes any command that is registered in extension, but this extension is not installed, they'll receive error:
`Unknown command ...`. Instead of showing this, it is better to tell them how to get access to this command.
In order to achieve this, implement a new logic in `extensibilityService` to find commands registered in all extensions. CLI searches npm (via npms API) for packages that have the keyword `nativescript:extension`.
After that, CLI gets the package.json of the latest version of each dependency (via registry.npmjs.org) and searches for nativescript key in it. In case it has nativescript key, CLI reads the commands value inside it. The value should be string array containg all commands registered in extension's bootstrap.
When user writes down a command that is not registered in CLI, we check if there's such command and in case it is found, CLI will instruc the user how to install the respective extension.
Same will happen in case user tries to access the help content of a command that is registered in extension.